+++ packages/voter-stake-registry-hooks/src/hooks/useCreatePosition.ts
@@ -9,6 +9,7 @@ import {
 import {
   Keypair,
   PublicKey,
+  SYSVAR_CLOCK_PUBKEY,
   SystemProgram,
   TransactionInstruction,
 } from "@solana/web3.js";
@@ -16,7 +17,13 @@ import { useAsync, useAsyncCallback } from "react-async-hook";
 import { useHeliumVsrState } from "../contexts/heliumVsrContext";
 import { HeliumVsrClient } from "../sdk/client";
 import { getRegistrarKey } from "../utils/getPositionKeys";
+import { SubDaoWithMeta } from "../sdk/types";
+import {
+  init as initHsd,
+  subDaoEpochInfoKey,
+} from "@helium/helium-sub-daos-sdk";
 
+const SECS_PER_DAY = 86400;
 export const useCreatePosition = () => {
   const { provider } = useHeliumVsrState();
   const { result: client } = useAsync(
@@ -29,15 +36,18 @@ export const useCreatePosition = () => {
       lockupKind = { cliff: {} },
       lockupPeriodsInDays,
       mint,
+      subDao,
       onInstructions,
     }: {
       amount: BN;
       lockupKind: any;
       lockupPeriodsInDays: number;
       mint: PublicKey;
+      subDao?: SubDaoWithMeta;
       // Instead of sending the transaction, let the caller decide
       onInstructions?: (
-        instructions: TransactionInstruction[], signers: Keypair[]
+        instructions: TransactionInstruction[],
+        signers: Keypair[]
       ) => Promise<void>;
     }) => {
       const isInvalid = !provider || !client;
@@ -46,9 +56,11 @@ export const useCreatePosition = () => {
       if (isInvalid) {
         throw new Error("Unable to Create Position, Invalid params");
       } else {
+        const hsdProgram = await initHsd(provider);
         const mintKeypair = Keypair.generate();
         const position = positionKey(mintKeypair.publicKey)[0];
         const instructions: TransactionInstruction[] = [];
+        const delegateInstructions: TransactionInstruction[] = [];
         const mintRent =
           await provider.connection.getMinimumBalanceForRentExemption(
             MintLayout.span
@@ -101,10 +113,49 @@ export const useCreatePosition = () => {
             .instruction()
         );
 
+        if (subDao) {
+          const clock = await provider.connection.getAccountInfo(
+            SYSVAR_CLOCK_PUBKEY
+          );
+          const unixTime = clock!.data.readBigInt64LE(8 * 4);
+          const registrarAcc = await client.program.account.registrar.fetch(
+            registrar
+          );
+          const currTs = Number(unixTime) + registrarAcc.timeOffset.toNumber();
+          const endTs = lockupPeriodsInDays * SECS_PER_DAY + currTs;
+          const [subDaoEpochInfo] = subDaoEpochInfoKey(subDao.pubkey, currTs);
+          const [endSubDaoEpochInfoKey] = subDaoEpochInfoKey(
+            subDao.pubkey,
+            endTs
+          );
+
+          delegateInstructions.push(
+            await hsdProgram.methods
+              .delegateV0()
+              .accounts({
+                position,
+                mint: mintKeypair.publicKey,
+                registrar,
+                subDao: subDao.pubkey,
+                dao: subDao.dao,
+                subDaoEpochInfo: subDaoEpochInfo,
+                closingTimeSubDaoEpochInfo: endSubDaoEpochInfoKey,
+                genesisEndSubDaoEpochInfo: endSubDaoEpochInfoKey,
+              })
+              .instruction()
+          );
+        }
+
         if (onInstructions) {
-          await onInstructions(instructions, [mintKeypair])
+          await onInstructions(
+            [...instructions, ...delegateInstructions],
+            [mintKeypair]
+          );
         } else {
           await sendInstructions(provider, instructions, [mintKeypair]);
+          if (delegateInstructions.length > 0) {
+            await sendInstructions(provider, delegateInstructions);
+          }
         }
       }
     }