+++ programs/hexboosting/src/instructions/start_boost_v0.rs
@@ -1,19 +1,19 @@
 use anchor_lang::prelude::*;
-use mobile_entity_manager::CarrierV0;
 
 use crate::{BoostConfigV0, BoostedHexV0};
+
+#[derive(AnchorSerialize, AnchorDeserialize, Clone, Default)]
+pub struct StartBoostArgsV0 {
+  pub start_ts: i64,
+}
 
 #[derive(Accounts)]
 pub struct StartBoostV0<'info> {
   pub start_authority: Signer<'info>,
-  pub boost_config: Box<Account<'info, BoostConfigV0>>,
   #[account(
-    has_one = hexboost_authority,
-    constraint = carrier.sub_dao == boost_config.sub_dao,
-    constraint = carrier.approved,
+    has_one = start_authority
   )]
-  pub carrier: Box<Account<'info, CarrierV0>>,
-  pub hexboost_authority: Signer<'info>,
+  pub boost_config: Box<Account<'info, BoostConfigV0>>,
   #[account(
     mut,
     has_one = boost_config,
@@ -21,9 +21,11 @@ pub struct StartBoostV0<'info> {
   pub boosted_hex: Box<Account<'info, BoostedHexV0>>,
 }
 
-pub fn handler(ctx: Context<StartBoostV0>) -> Result<()> {
-  let now = Clock::get()?.unix_timestamp;
-  ctx.accounts.boosted_hex.start_ts = now;
+pub fn handler(ctx: Context<StartBoostV0>, args: StartBoostArgsV0) -> Result<()> {
+  require_eq!(ctx.accounts.boosted_hex.start_ts, 0);
+  require_gt!(args.start_ts, 0);
+
+  ctx.accounts.boosted_hex.start_ts = args.start_ts;
 
   Ok(())
 }