+++ packages/faucet-service/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.29](https://github.com/helium/helium-program-libary/compare/v0.6.28...v0.6.29) (2024-02-06)
+
+**Note:** Version bump only for package @helium/faucet-service
+
+
+
+
+
+## [0.6.28](https://github.com/helium/helium-program-libary/compare/v0.6.27...v0.6.28) (2024-02-06)
+
+**Note:** Version bump only for package @helium/faucet-service
+
+
+
+
+
+## [0.6.27](https://github.com/helium/helium-program-libary/compare/v0.6.25...v0.6.27) (2024-02-01)
+
+**Note:** Version bump only for package @helium/faucet-service
+
+
+
+
+
+## [0.6.26](https://github.com/helium/helium-program-libary/compare/v0.6.5...v0.6.26) (2024-01-31)
+
+**Note:** Version bump only for package @helium/faucet-service
+
+
+
+
+
+## [0.6.25](https://github.com/helium/helium-program-libary/compare/v0.6.24...v0.6.25) (2024-01-29)
+
+**Note:** Version bump only for package @helium/faucet-service
+
+
+
+
+
+## [0.6.24](https://github.com/helium/helium-program-libary/compare/v0.6.23...v0.6.24) (2024-01-29)
+
+**Note:** Version bump only for package @helium/faucet-service
+
+
+
+
+
 ## [0.6.23](https://github.com/helium/helium-program-libary/compare/v0.6.20...v0.6.23) (2024-01-20)
 
 **Note:** Version bump only for package @helium/faucet-service