+++ packages/helium-entity-manager-sdk/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.29](https://github.com/helium/helium-program-libary/compare/v0.6.28...v0.6.29) (2024-02-06)
+
+**Note:** Version bump only for package @helium/helium-entity-manager-sdk
+
+
+
+
+
+## [0.6.28](https://github.com/helium/helium-program-libary/compare/v0.6.27...v0.6.28) (2024-02-06)
+
+
+### Features
+
+* **#559:** Use pyth oracle for mobile onboarding ([#566](https://github.com/helium/helium-program-libary/issues/566)) ([1b751e9](https://github.com/helium/helium-program-libary/commit/1b751e9a1aabbbe851ce828c4b2dfff46d18615a)), closes [#559](https://github.com/helium/helium-program-libary/issues/559)
+
+
+
+
+
+## [0.6.27](https://github.com/helium/helium-program-libary/compare/v0.6.25...v0.6.27) (2024-02-01)
+
+**Note:** Version bump only for package @helium/helium-entity-manager-sdk
+
+
+
+
+
+## [0.6.26](https://github.com/helium/helium-program-libary/compare/v0.6.5...v0.6.26) (2024-01-31)
+
+
+### Features
+
+* **#494:** Implement HIP-96 wifi onboarding fees ([#506](https://github.com/helium/helium-program-libary/issues/506)) ([50471c3](https://github.com/helium/helium-program-libary/commit/50471c33d06ba7f55e605ed7dbba02119af5e3cd)), closes [#494](https://github.com/helium/helium-program-libary/issues/494)
+
+
+
+
+
+## [0.6.25](https://github.com/helium/helium-program-libary/compare/v0.6.24...v0.6.25) (2024-01-29)
+
+**Note:** Version bump only for package @helium/helium-entity-manager-sdk
+
+
+
+
+
+## [0.6.24](https://github.com/helium/helium-program-libary/compare/v0.6.23...v0.6.24) (2024-01-29)
+
+**Note:** Version bump only for package @helium/helium-entity-manager-sdk
+
+
+
+
+
 ## [0.6.23](https://github.com/helium/helium-program-libary/compare/v0.6.20...v0.6.23) (2024-01-20)
 
 **Note:** Version bump only for package @helium/helium-entity-manager-sdk